add --clean-up-orphaned-replica-sets

+++ kube/patch.sh
@@ -3,14 +3,17 @@
 set -e
 set -u
 
+CLEAN_UP_ORPHANED_REPLICA_SETS='--clean-up-orphaned-replica-sets'
 usage() {
-  echo -e "Usage: $0 <deployment|statefulset> <name>\n"
+  echo -e "Usage: $0 <deployment|statefulset> <name> <${CLEAN_UP_ORPHANED_REPLICA_SETS}>\n"
 }
 
 if [  $# -le 1 ]; then
   usage
   exit 1
 fi
+
+SHOULD_CLEAN_UP=${3:-}
 
 # Override to use a different Docker image name for the sidecar.
 export SIDECAR_IMAGE_NAME=${SIDECAR_IMAGE_NAME:-'gcr.io/stackdriver-prometheus/stackdriver-prometheus-sidecar'}
@@ -37,5 +40,7 @@ spec:
         - name: ${DATA_VOLUME}
           mountPath: ${DATA_DIR}
 "
-# Delete the replica sets from the old deployment.
-kubectl -n "${KUBE_NAMESPACE}" get rs | grep "$2-" | awk '{print $1}' | xargs kubectl delete -n "${KUBE_NAMESPACE}" rs
+if [[ "${SHOULD_CLEAN_UP}" == "${CLEAN_UP_ORPHANED_REPLICA_SETS}" ]]; then
+  # Delete the replica sets from the old deployment.
+  kubectl -n "${KUBE_NAMESPACE}" get rs | grep "$2-" | awk '{print $1}' | xargs kubectl delete -n "${KUBE_NAMESPACE}" rs
+fi
